+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [4.0.0] - 2020-05-18
+### Removed
+- Remove Insite ID Store adapter
+### Added
+- New config option for converting Workday fields into ID Broker 'groups'
+
 ## [3.3.1] - 2020-05-12
 ### Fixed
 - Application logs from console scripts sent to stdout
@@ -171,7 +177,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - First release.
 
-[Unreleased]: https://github.com/silinternational/idp-id-sync/compare/3.3.1...develop
+[Unreleased]: https://github.com/silinternational/idp-id-sync/compare/4.0.0...develop
+[4.0.0]: https://github.com/silinternational/idp-id-sync/compare/3.3.1...4.0.0
 [3.3.1]: https://github.com/silinternational/idp-id-sync/compare/3.3.0...3.3.1
 [3.3.0]: https://github.com/silinternational/idp-id-sync/compare/3.2.0...3.3.0
 [3.2.0]: https://github.com/silinternational/idp-id-sync/compare/3.1.0...3.2.0